+++ cdc/metrics/metrics_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package owner
+package metrics
 
 import (
 	"testing"
@@ -20,10 +20,12 @@ import (
 )
 
 func TestLagBucket(t *testing.T) {
-	buckets := lagBucket()
-	require.Equal(t, 40, len(buckets))
+	buckets := LagBucket()
+	require.Equal(t, 64, len(buckets))
 	require.Equal(t, 0.5, buckets[0])
 	require.Equal(t, 1.0, buckets[1])
 	require.Equal(t, 21.0, buckets[30])
-	require.Equal(t, float64(10752), buckets[39])
+	require.Equal(t, 900.0, buckets[50])
+	require.Equal(t, 4000.0, buckets[60])
+	require.Equal(t, float64(32000), buckets[63])
 }
