+++ cdc/owner/metrics.go
@@ -16,6 +16,7 @@ package owner
 import (
 	"time"
 
+	"github.com/pingcap/tiflow/cdc/metrics"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -56,7 +57,7 @@ var (
 			Subsystem: "owner",
 			Name:      "checkpoint_lag_histogram",
 			Help:      "checkpoint lag histogram of changefeeds",
-			Buckets:   lagBucket(),
+			Buckets:   metrics.LagBucket(),
 		}, []string{"namespace", "changefeed"})
 
 	changefeedResolvedTsGauge = prometheus.NewGaugeVec(
@@ -80,7 +81,7 @@ var (
 			Subsystem: "owner",
 			Name:      "resolved_ts_lag_histogram",
 			Help:      "resolved_ts lag histogram of changefeeds",
-			Buckets:   lagBucket(),
+			Buckets:   metrics.LagBucket(),
 		}, []string{"namespace", "changefeed"})
 
 	ownershipCounter = prometheus.NewCounter(
@@ -150,17 +151,4 @@ func InitMetrics(registry *prometheus.Registry) {
 	registry.MustRegister(changefeedTickDuration)
 	registry.MustRegister(changefeedCloseDuration)
 	registry.MustRegister(changefeedStartTimeGauge)
-}
-
-// lagBucket returns the lag buckets for prometheus metric
-// 10 seconds is the reasonable LAG for most cases,
-// for prometheus histogram_quantile func,
-// we use small bucket distance to do accurate approximation
-func lagBucket() []float64 {
-	buckets := prometheus.LinearBuckets(0.5, 0.5, 20)
-	buckets = append(buckets, prometheus.LinearBuckets(11, 1, 10)...)
-	buckets = append(buckets, prometheus.LinearBuckets(21, 30, 20)...)
-	buckets = append(buckets, prometheus.LinearBuckets(900, 300, 10)...)
-	buckets = append(buckets, prometheus.ExponentialBuckets(4000, 2, 4)...)
-	return buckets
 }
