+++ cdc/sink/metrics/tablesink/metrics.go
@@ -14,6 +14,7 @@
 package tablesink
 
 import (
+	"github.com/pingcap/tiflow/cdc/metrics"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -32,20 +33,11 @@ var TableSinkFlushLagDuration = prometheus.NewHistogramVec(
 		Subsystem: "sink",
 		Name:      "flush_lag_histogram",
 		Help:      "flush lag histogram of rows that are processed by table sink",
-		Buckets:   lagBucket(),
+		Buckets:   metrics.LagBucket(),
 	}, []string{"namespace", "changefeed"})
 
 // InitMetrics registers all metrics in this file.
 func InitMetrics(registry *prometheus.Registry) {
 	registry.MustRegister(TotalRowsCountCounter)
 	registry.MustRegister(TableSinkFlushLagDuration)
-}
-
-func lagBucket() []float64 {
-	buckets := prometheus.LinearBuckets(0.5, 0.5, 20)
-	buckets = append(buckets, prometheus.LinearBuckets(11, 1, 10)...)
-	buckets = append(buckets, prometheus.LinearBuckets(21, 30, 20)...)
-	buckets = append(buckets, prometheus.LinearBuckets(900, 300, 10)...)
-	buckets = append(buckets, prometheus.ExponentialBuckets(4000, 2, 4)...)
-	return buckets
 }
